current draw fixes
changed current draw variable type from into to float

+++ Battery_Box_Board/Constants.h
@@ -54,7 +54,7 @@ bool sdFailure;
 float soc;
 int dcl;
 float cellLowV;
-int currentDraw;
+float currentDraw;
 int cellHighTemp;
 bool currentFail;
 bool voltageFail;
